[DC-1930] Migrate constants and unit tests
