--- data_steward/constants/retraction/retract_deactivated_pids.py
@@ -1,8 +1,0 @@
-START_DATE = 'start_date'
-START_DATETIME = 'start_datetime'
-END_DATE = 'end_date'
-END_DATETIME = 'end_datetime'
-DATE = 'date'
-DATETIME = 'datetime'
-
-DEACTIVATED_PARTICIPANTS = '_deactivated_participants'
